requirements.txt
python-docx
#Something regarding LibreOffice
